Refactor TopEnvFrag
TopEnvFrag was very confusing, because even though it used to carry a
ModuleEnv, many elements of that ModuleEnv (source map, synthesis candidates, ...)
would be expected to not include any imported values! I lost some time
thanks to this unstated invariant, so I figured it's a good time to
clean it up!

With this change `TopEnvFrag` directly mimics the structure of `Env`,
and includes top emissions separately from module-level emissions. And
module-level emissions are no longer represented by a `ModuleEnv`, but
by a `ModuleEmissions` ADT that explicitly states the relevant
invariants.