Add a concept of "sub fragment" and use it to optimize Imp's Dest emissions
DestEmissions are a very rich structure that is a bit complicated to
extend. At the same time in most places we emit only a tiny component of
the whole emissions, which means that we still have to pay for no-op
extensions of the other parts.

In this change I add a concept of "sub fragment" that effectively
represents a specialized update to the inplace environment and
emissions. Thanks to inlining and specialization GHC is able to elide
the unnecessary extensions and make the code both faster and have a much
less noisy Core representation.