Make a dedicated expression language for synthesized dictionaries
Previously we used Blocks for synthesized dictionaries. This meant that
synthesis would emit decls, which made it hard to have synthesized dictionaries
(like `Ix`) in types (like those producing table types). With this change,
synthesis expressions can appear in atoms.

Why is this ok? There are two reasons we distinguish atoms from blocks:
  1. We like to do equality checking syntactically. Atoms are normalized but
     blocks aren't.
  2. We want to avoid duplicating effects and runtime work. Atoms are
     effect-free and runtime-free so we can freely inline them.

But synthesis only uses two operations: instantiation and superclass projection.
These operations don't have any effects or runtime work. And even when
dictionaries appear in types, we won't be using the values for checking equality
because we assume they're uniquely determined by their type. This all means
there's no harm in putting synthesis expressions in Atoms.

It also seems to be a performance improvement: it reduces allocations by 10% in
the kernel regression example.

It also solves the problem that synthesizing dictionaries would generate code
for non-function methods. For example, writing `[1,2,3] + [1,2,3]` would create
an array of zeros because of the `zero` method of `Add`! With this change, decls
are only emitted when you actually project the method (and only the decls for
that method). So we don't need the thunked method workaround anymore.