Perform incremental outputs in all but HTML mode
Right now we wait until a script is fully evaluated before we even
attempt to produce any output. It's confusing and bad UX, because it
almost looks like a hang until a wall of text appears at the very end.
This reorganizes the executable and top level so that we print the
evaluated source blocks interactively, making Dex seem much more
responsive.

Also, restore support for aborting on errors (with the new flag enabling
the aborts and the default being to continue until the end of the script).