Simplify binary lambdas with intermediate decls
We used to make a simplifying assumption that the binary lambda
arguments to effect handlers have no decls in between the first and
second lambda binder. However, it's been easy to violate that
assumption, because inference likes to insert decls while unpacking
patterns. This meant that accumulating over a monoid such as
`AddMonoid Complex` caused a simplification error, because of `Complex`
argument unpacking.

This change relaxes the constraint slightly, to allow for the
problematic decls. However, it's still not a complete solution as it
assumes that the second binder and the eventual reconstruction don't
depend on those decls. This seems sufficient for now and we can always
revise that in the future.