Make WriterT1 use HoistableState instead of FallibleMonoid.
With FallibleMonoid, hoisting the written state was all or nothing. With
HoistableState, the implementor of the state type gets to choose how hoisting
should happen. We need this in CheapReduction, where we want to hoist those
dict types that are possible to hoist even if there are others that can't be
hoisted.