Refactor Blocks to terminate in Atoms rather than Exprs.
It turns out that this was already the case, except for the
consequences of `inlineLastDecl`.  Moved that out of Builder and into
the two consumers that seem to depend upon it, namely PPrint.hs (of a
chain of for expressions?) and Serialize.hs (which has a fast
path guarded by pattern-matching for trivial string tables).