Specialize defaults for integer literals
The previous implementation of defaults was very expressive, but it was
all for nothing! The only place where we actually use that is for
integer literals, so the list of pairs can be easily replaced with a
`NameSet` that's much cheaper to deal with!

This single refactor speeds up an inference-heavy workload by 15% and
saves us 20% of allocations.