Optimize hoisting inference state
On one of the benchmarks I tried this function took over half the time
required for inference. It's really important!

This change optimizes it a bit. It iterates over RNests, making it tail
recursive. This also adds a bunch of fast paths that skip consistency
checks. Of course, all of those are disabled in a debug build that uses
100% safe APIs.