Add strictness annotations to let GHC unpack free vars cache
The new data structures that cache and delay various pieces of state
during `hoistInfStateRec` are meant to be transient and can be splatted
out into the argument list by the compiler. But, GHC can only do it if
it knows those arguments are strict in every branch, which they weren't.
However, we don't expect any divergence from those, so it should be safe
to mark them as strict.