Add a GitHub action for continuous benchmarking
We've put in a lot of effort into making Dex faster in the recent months
and it would be a waste to have it all slip away. So, to prevent that,
I've set up a simple system for continuous performance tracking. This
includes three parts:
1. I created an orphan branch named `performance-data`, which we'll use
as a simple CSV-based database of performance measurements. I ran a few
benchmarks using the new `benchmarks/continuous.py` script to
pre-populate it.
2. On the `gh-pages` branch, there's a new `performance.html` file,
now accessible via [this URL](https://google-research.github.io/dex-lang/performance.html).
It pulls the data from the `performance-data` branch and displays it as
a series of plots showing changes in total allocation and end-to-end
execution times of a few of our examples.
3. This commit checks in the file used to perform benchmarks, and adds a
GitHub action that will run it every time we push to `main`.

While total allocation numbers are very stable, at this point you might
be worried that any time estimates we get from the free VMs that execute
GitHub actions will be too noisy to provide any good signal. And if we
did it naively, you wouldn't be wrong! To make the drift between
different machines lower we use the techniques outlined in
[this post](https://labs.quansight.org/blog/2021/08/github-actions-benchmarks/).
The idea is to report _relative change_ in execution time compared to
baseline. Most importantly, any time we want to evaluate a new commit,
_we rebenchmark the baseline_ to adjust for the differences between
machines. To minimize noise on shorter time scales, we interleave the
evaluations of baseline compiler with those using newer versions.