Add a `Post n` index set to represent the fenceposts of `n`.
This gives a neat way to handle "fencepost" problems, where you have two sets of
entities whose sizes differ by one, e.g. a sequence of states and the
transitions between those states, or the characters in a string and the set of
prefixes of the string. Usually, in a world where indices are integers, you have
to decide on a convention for how to map the "post" indices to the "segment"
indices and it's a common source of off-by-one errors. In Dex, we can just use
separate index sets for the posts and the segments, with functions that map
between them. Each segment has a previous post and a next post. Each post
*might* have a previous or next segment.

This also introduces the `NonEmpty` constraint on index sets, but for now the
only instance is for `Post n`.