Add support for monomorphic literals to speed up jax2dex
Jaxprs translated to Dex have plenty of literals scattered around,
causing us to repeat `FromInteger` dict synthesis all over the place.
This is of course unnecessary, because JAX has full type information.
With this change literals spelled up as `(%monoLit 1234)` do not
participate in the usual polymorphic elaboration, but are always typed
as 32-bit integers.

Note that having lots of literals is very unidiomatic for Dex, so this
is generally not an issue for human-written programs.