Allow dex web to be used without a network connection.
The issue was that the dynamic content rendering loop was only started
after KaTeX was loaded, in the `onload` property.  But naturally
enough that trigger doesn't run if the load fails, as when there is no
network.  Added starting the same loop `onerror` as well, and tweaked
the Dex-specific Javascript to survive KaTeX being unavailable (by
just not rendering any LaTeX with it).

Fixes #610.