Refactor and optimize generic traversals
Soon enough we'll need to start adding optimization passes, and those
will have to be implemented on top of generic traversals. I glanced over
the current implementation, and while it was nice conceptually, it
didn't expose too many optimization opportunities for GHC.

In this commit, I make the generic traversal monad concrete, and only
allow the different traversals to customize the type of state carried by
the monad. Of course each traversal can also customize both
`traverseExpr` and `traverseAtom`, with type of the monadic state acting
as a selector for the appropriate typeclass methods. While this loses
flexibility compared to the old approach (which could run with an
arbitrary stack), it should help us avoid unnecessary slowdowns in the
compiler.

Finally, I also use other tricks like the tail-recursive decl traversal
or monad eta-expansion that we already implement in Simplification.