Add an early optimization pass that unrolls trivial loops
This adds a transform that eliminates loops that use `Fin 0` and `Fin 1`
as their index set. In the future we can extend it to user-defined index
sets as well, by looking up the `size` in the `Ix` dictionary. But since
those are still undergoing some changes, I decided to hard-code the
builtin types for now.

Note that it only runs _before simplification_, meaning that it's
complexity is proportional to the source program size. Since those are
generally small, it should be very cheap to apply. It misses some
opportunities for further optimizations that might get exposed once
polymorphic functions get instantiated, but we can always repeat this
pass in a later optimization pipeline.