Rewrite type inference for table literals.
Previously, it would attempt to match an incoming type for a table
literal only if the index set of that type was closed, which prevented
things like `[0.0, 1.0]:(Bool=>Float)`.  (With, might I add, a rather
unhelpful error message suggesting the user add a type annotation to
this literal.)

Now, inferring a table literal with requested type n=>a goes like this:
- Try to synthesize an Ix dictionary for n
- Obtain the size method thereof and attempt to execute it in the
  interpreter
- Confirm that the size equals the number of elements in the table
  literal
- Type-check each element according to the requested type
- If any of the first three steps above fail, fall back to typing the
  literal as `Fin k` where k is its length.

This is better because we use the full power of the interpreter for
trying to figure out the size of the requested index set, which
includes looking up variables and doing index arithmetic, and thus
fixes #563.

It is also better because it compares the size of the requested index
set with the length of the literal before trying to syntehsize the
full list of indices, and thus fixes #912.