Add support for the XLA:CPU calling convention in export
Dispatch times of Dex computations launched from JAX is pretty high,
because at the moment we use a Python-based trampoline that adjusts the
calling conventions to match. In this patch I add support for the
XLA calling convention to our codegen, which lets us eliminate the
Python indirection and have a fully native jump.