Add a prototype of an incremental lowering path
The jump from Core to Imp (implemented in `Imp.hs`) is huge! We have to
do the AoS/SoA conversion, bufferize our code, do destination passing to
elide copies (which apparently we no longer do since safer names!),
assign buffers to memory spaces, etc etc. This is a lot, and once you do
so much in a single pass it is easy to make mistakes, or suboptimal
decisions.

To solve that problem I want us to have an intermediate "abstract
machine" IR for Dex, which is where most of those operations would be
performed. And it would also be the level where many optimizations would
be performed. This includes loop splitting, when an index set is a sum
type, which can be generalized to tiling! And, since we can talk about
mutable buffers, we won't need _any_ custom special primitives ops for that!

This change is a tiny first step in that direction. The code is
effectively dead, but it's good enough to explore some simple ideas. As
a follow-up, I will want to add a flag that conditionally enables this
progressive lowering feature, and then we can start fleshing out the
support in codegen etc.