Remove built-in index ranges and reimplement them in user space.
This makes `inject` a user-defined function too. It's based on a new
user-defined `Subset` class, with a total `inject` method and a partial
`project` method. It ought to be transitive and reflexive, but that isn't
captured yet because of limitations in dict synthesis.

 I also removed ranges with both upper and lower bounds. They're complicated,
 they have strange semantics when the lower bound is higher than the upper
 bound, and we haven't found a need for them in practice. You can always
 simulate them using two one-sided slices, and that has the additional benefit
 that it forces you to handle the case where the bounds cross.

This breaks some AD tests because we don't have tangent spaces for the new
user-defined index range types. I'll fix that next.