Tangent types for "newtype" wrappers.
If we're asked for the tangent type of a user-defined ADT with a newtype wrapper
structure (a single data constructor with a single argument) we use the tangent
type of the underlying wrapped type. This doesn't fix the problem of doing AD on
general user-defined types, but it handles cases like the new `Range*` types.