Proposal to change the syntax for "type-annotate expression" to use `::`, not `:`.
Why?  The `:` already means "type-annotate binder", and in particular
serves as the marker that distinguishers a binder from a variable
reference.  For instance,

  foo : (Fin 3) => Float

parses as a pi type for an array whose indices are named `foo`, but
whose elements happen not to depend on the index, even though it is in
scope.  In the situation where the user wants to write a reference
to the variable `foo` and annotate it with the type `(Fin 3) =>
Float`, this change allows them to write

  foo :: (Fin 3) => Float

instead of the unintuitive (to me, at least)

  (foo) : (Fin 3) => Float

that is required by the status quo.

This PR also tweaks the parse rule for `decl` to commit only when it
sees the `=` sign, rather than committing eagerly on `:`.  That way,
we can keep the behavior that

  foo : (Fin 3) => Float = stuff

parses as binding `foo` to `stuff` while annotating `foo` with the
type `(Fin 3) => Float` while simultaneously allowing a line beginning
with

  foo : (Fin 3) => Float

to parse as a pi type instead of trying to parse as a binding of `foo`
and failing.

Committing later is presumably something of a performance regression
for the parser, but I think it only matters in this case, namely
trying to parse

  foo : (Fin 3) => Float

as a binding, then giving up for lack of an `=` sign and re-parsing it
as a pi type.

Fixes #933.