Add tutorial data download commands to Makefile
The data is downloaded precisely as a dependency to the tutorial test
case. Naturallly, this requires an internet connection to succeed, but
one may also place the data files in the expected location manually
(e.g. as GitHub Actions currently does).

Also adds a phony `tutorial-data` target for easy downloading (e.g. if
we want to update the GHA workflows to use this) and adjusts the clean
target appropriately.