Use GHA concurrency groups, fix caching MNIST data
GitHub Actions workflows on pull requests are not normally canceled when
a newer commit comes in on the same branch. This wastes CI minutes and
delays PRs. The three CI workflows now use GHA's concurrency groups
feature to cancel stale workflows on the same pull request. Note that
this will not affect workflows on main after merging.

This commit also fixes the unconditional download of the Fashion MNIST
data set by delegating to the Makefile and removing the workflow step.