Do algebraic simplification only on Nat vars, avoiding the index/ordinal confusion.
Previously, polynomial terms in the Algebra module could contain variables
either of type `Nat` (originally `Int`) or an arbitrary index type. We
implicitly identified index-typed variables with their ordinals. But we baked in
incorrect assumptions about the ordinal/index relationship. The situation would
be even worse with #935 is merged, because then user-defined index could diverge
even further from those assumptions. For example, with user-defined index sets
like `RangeFrom`, ordinal/from_ordinal become wrapping/unwrapping the newtype
constructor. #935 currently adds that special case as a hack, but we want to
avoid that.

We could try to teach Algebra about index sets, but sometimes there can be
variables appearing in a polynomial expression that aren't bound directly by an
index binder, so that gets tricky. Instead, this change updates, Algebra
polynomials to work *only* with Nat-typed variables. The module basically
becomes a tool for attempting to find a closed form for
`\sum_{i=0}^(lim-1) body`, where `i`, `lim`, and `body` all have type `Nat`.

When we call Alebgra from Imp, we give it an abstraction whose argument is a
`Nat`, which we then convert to an ordinal, which will then presumably be
converted back to a `Nat` in the user-defined implementation of `size` or
something. We rely on simplification to cancel out this conversion from nat to
index and back to nat, so that we end up with just primitive operations on nats,
`iadd`, `imul`, etc.

This cancellation will fail if simplification can't tell that `ordinal` and
`unafe_from_ordinal` are inverses of each other. This happens with `Fin (a & b)`,
for example (though that didn't work before either). One solution would be to
avoid inlining `ordinal` and `unsafe_from_ordinal` before Imp. Then we could
cancel them out directly, rather than trying to cancel out their
implementations.