Fix a bug in destination passing implementation
Destination passing tries to lift the writes to the destination for the
block result up to the places where the written values are defined. But,
not all block results have to be defined in the same block! The old
implementation would fail to insert the necessary writes, leaving us
with uninitialized dests.