Add a Read-like interface for parsing Strings into Dex values
... with an initial implementation supporting `Float`s, with parsing
implemented via the `strtof` C function. Obviously that won't fly on the
GPU, but it's not like `String`s work great in that setting, so it
should be good enough for now.

I wanted to call this interface `Read` to match the Haskell
implementation, but it turns out it's already taken for the `Read`
effect. I think there's little reason to even keep that effect visible
in the user space, since one might just pass the value that can be read
in place of the reference it can be read from...