Add support for declaring custom function linearizations
This change adds a new syntax for
```
custom-linearization var expr
```
which declares `expr` to be the expression defining the custom
linearization of function defined by `var`. If the type of `var` is
`a -> b`, then `expr` should have type `a -> (b, a -> b)`. I didn't use
linear arrows, because linearity is a bit sketchy in the surface
language at the moment.

It is however still limited, and is mostly meant to be a prototype
implementation to figure out most annoyances. I will make it more
general in following changes (adding support for nary functions and
polymorphism).

---

This required quite a bit of care around simplification, because now the
identities (names) of functions have meaningful effects on AD, so we
can't just inline them all over the place. As such, we now refrain from
inlining bodies of functions defined at the top level. They only get
inlined when they appear on the lhs of an application, except for when
we're under `linearize` and we know that the function has a custom rule
defined.

Note that this inlining now sometimes has to go through multiple hops.
Consider the following program:
```
f = \x. ...
g = f
g 1.0
```
Before, when we began analyzing the expression `g 1.0`, the environment
roughly looked like this:
```
f @> \x. ...
g @> \x. ...  -- inlined body of f
```
However, after this change, `f` will not get inlined eagerly into the
definition of `g`, and the env will look as follows:
```
f @> \x. ...
g @> f
```
meaning that inlining of `g` into `g 1.0` requires us to go through a
chain of lookups.

The impact of that change on performance is unclear, but I would expect
it to be a net positive. We might end up doing more lookups, but they're
generally cheap, and we generally don't write long chains of function
aliases. At the same time, we actually can gain quite a bit by inlining
less eagerly. If we were to serialize the environment after evaluating
the example program ablve, the new one would have a much more compact
representation than the old.