Map BUILTIN_LIBRARIES in --lib-path to the standard library path
`--lib-path` is useful to extend the search path with custom libraries,
but it also removes the default search path, unless it is specified
explicitly. After this patch one can use a special `BUILTIN_LIBRARIES`
placeholder in the directory list, which will get replaced with the
directory in which the standard libraries have been installed.