Allow custom linearizations of polymorphic functions
Unfortunately, this has a bit of a limited utility, because full
polymorphism is still not allowed. That's because the tangent type
mapping has no user-space definition, so there's no way to spell out a
function of type `Tangent a -> Tangent a` (which is the type of
linearization of `a -> a`). But we can at least handle shape polymorphic
function.