Add support for symbolic representation of zero tangents in custom linearization rules
When a `custom-linearization-symbolic` directive is used to declare a
custom linearization rule, the linearized function is expected to take
`SymbolicTangent a` as the type of tangents of `a`. This makes it
possible to branch based on the activity of function arguments and use
that information to choose the most efficient linearization rule.