Implement dead code elimination
This change reimplements dead code elimination after safer names. I
think that it's the first of the passes that utilize information about
binder usage, so it was a good exercise that will make the
implementation of other similar passes (e.g. inlining) easier.

Thanks to the way we represent bindings in Dex Core, DCE is fundamentally
about traversing the whole IR in search of blocks, and then performing
DCE on every block. I thought about using a generic traversal for that
purpose, but ultimately decided against it. It wasn't a good match,
because it is written with forward-traversals in mind, and requires us
to carry a substitution, which is not necessary in the case of DCE.

The implementation strategy I ended up with is very similar to the way
we generate code for substitutions and free variable computation. Which
makes sense, because DCE also wants to compute free variables. To avoid
repeated traversals of subtrees, we fuse this computation with the
block-searching-traversal. Note that the hoisting component is a bit
unsafe, but we verify the invariants in debug mode (by paying for the
repeated traversals).

I also verified that all the generics get inlined and optimized away, so
this approach ends up yielding code similar to one we would get by
writing explicit `case` expressions over our IR. However, this approach has
the benefit of automatically adapting to any changes in the Core IR.