Avoid specializing on the "data" components of types.
The idea is that we don't want to specialize so aggressively that `Fin 4` and
`Fin 5` yield different specializations. In both cases, we only want to
specialize at `Fin n` "for all n". For now, we've special-cased this to `Fin`,
but we should be able to generalize to any type parameter that has a runtime
representation (i.e. anything we consider "data").