Make sure that variables defined outside of the block are placed
Normally, when variables to be placed in a destination are defined
within the decls of the block, the `Place` operations get emitted during
the decl traversal. However, we didn't handle the case when they're
defined outside of the block!

This also switches destination propagation to use a RawName-based map,
for efficiency.