Make output destination allocation explicit outside of Imp
Right now, Imp is responsible for creating the allocation for the result
of the whole (lowered) compiled block, which means that it doesn't
participate in the destination passing implemented during Lowering. That
usually causes us to include unnecessary copies in our compiled code.
To work around it, we make the abstraction of the output dest explicit
during lowering.