Infer implicit arguments for jax.jit(dex.eval()).
To wit, we already had code in `dex_call_abstract_eval_with_shape` for
inferring implicit arguments; this change just reuses it in
`dex_apply_lowering` to construct appropriate MLIR constants to pass
as arguments to the compiled function.

The catch is that this whole system relies heavily on the
restrictedness of types that can travel across the Jax-Dex boundary,
to wit, only scalars and rectangular arrays; and on the fact that the
only allowed implicit arguments are the sizes of the dynamic
dimensions of said rectangular arrays.

However, that restriction is already present in the status quo before
this change.