Add support for partial destination passing
Our destination passing implementation used to only work with full
destinations. Which is generally fine, for as long as the result atoms
of blocks mention bare variables. However, many functions such as
`yieldAccum` end up with a projection in the result, which completely
trips up our method and results in many unnecessary copies.

In this patch I add a "projected dest" and make sure that we can e.g.
inherit a destination for an accumulator from the result of its
surrounding for loop.