Add a newtype constructor
The Core IR has quite a bit of redundancy in its representations.
Records, one constructor ADTs and products are all... just products,
with different types. And we should represent them that way!

This change adds a "newtype" atom constructor that can take a single
value payload to construct another type. At the moment it is only used
for `Fin` values (letting us remove the `FinVal` constructor). `Nat`s
(as a type different from Word32) are the next step.