A reasoned proposal for operator precedence in Dex.
As a side-effect of working on the parser reorganization #1013, I
spent some effort on thinking about operator precedence.  The result
is a proposal to adjust Dex's operator precedence table to the
structure and for the reasons given in this document.

The parser reorganization partially implements this propsal, because
it makes various formerly-special syntactic entities (like the equals
character `=` and the pipe character `|`) be operators subject to
precedence rules.

However, this proposal goes beyond the changes in the parser reorg.
Mainly, I propose defining explicit (and distinct) precedences for a
large number of oprators that are currently defined in user-space in
the prelude and not mentioned in the parser at all; and I also propose
changing the precedence of several extant operators relative to each
other.