Use newtypes to represent ADTs
ADTs now desugar to newtypes over sums of (dependent) products. Of
course, when there's a single constructor we elide the sum. Similarly,
when a constructor has a single field, we elide the product. Finally,
constructors with no dependent fields do not use dependent products in
their representation.

This patch already seems to simplify the implementation quite a bit, but
there are still a few improvements left on the table:
- `DataConRef` is dead and should be removed
- `Alt`s don't need binder nests anymore
- `DataConDef`s shouldn't need to carry the binder nest, just the
  representation type
- `SumAsProd` type is now always a sum type

And beyond the simplifications, it also generalizes our support for
dependent ADTs! We used to only allow dependent fields in
single-constructor types, but this requirement is now lifted (with the
implementation of dependent pair types doing all the heavy lifting).