Occurrence analysis data structures and unit tests
This change adds a series of internal representations for binding
occurrence information and transformations between them.  The
proximate goal is to support inlining array-valued bindings into loop
bodies, where safe; but in the future some evolution of these could
perhaps be used for assessing e.g., when a reference is written to
exactly once in each cell so that it can be replaced with a producer
loop.

The occurrence analysis structures here are coded to be able to see
through algebraic index types, to detect that e.g.
  for i. xs.(Left i)
  for j. xs.(Right j)
in fact reads `xs` only once per element.

This change also adds a `make watch-unit-tests` target that asks
`stack` to rerun the Haskell unit tests in response to source file
edits.

A couple open issues with this change I want to record for posterity:

- These structures only keep track of work (e.g., FLOP) preservation,
  not locality.  Inlining a locality-sensitive producer into
  transposition could be problematic; if we care about preventing
  that, these structures would need to be updated to keep some notion
  of locality in mind.

- There is a known precision problem: the pipeline is conservative
  about the number of dimensions of the binding that need to be
  apparent for inlining to be safe.  For instance, if the use is
    for j. xs.j.j
  then it's not work-preserving to inline `xs` if its definition
  cannot be immediately cancelled against one level of indexing.
  However, the analysis will conservatively assume two levels of
  indexing are required.