Perform loop invariant code motion (and DCE) after lowering
Most importantly, to lift allocs of arrays that have stable (but not
necessarily statically known) shapes out of the hot inner loops. This
dramatically improves the performance (40x !) of the `conv` benchmark
when it is not specialized on the `size` parameter (you can replace
the `3` by `sum [3]` and see for yourself).