Account for the cost of TabCon emitted after loop unrolling
Unrolling loops with large TabCons is a bad idea and #1009 was
supposed to prevent that. But I forgot to account for the TabCon that
we emit after unrolling a loop, so `for i:(Fin 10000) j:(Fin 4). ordinal j`
ended up unrolling the `[0, 1, 2, 3]` literal introduced by unrolling
the `j` loop 10000 times.