Remove `abstractPtrLiterals` since pointer literals are now handled at link time.
I had to tweak the prelude a bit because the null pointer constructed by casting
from the zero integer isn't getting linked properly. It's a justifiable change
in its own right because it's silly to treat the null pointer as a link-time
constant. But I don't understand why it caused a problem and I should figure it
out.