Use ObjectLinkingLayer on macOS
ObjectLinkingLayer is an OrcJIT interface to the new JITLink LLVM
library. It's supposed to be able to handle much more than the old
RuntimeDyld system, like relocations for code using the small
code model (see [this](https://llvm.org/docs/JITLink.html#runtimedyld)
for a list of restriction).

Unfortunately, we're still on LLVM 12 and at that time the support for
ELF relations in ObjectLinkingLayer (which are used on Linux) were rather
limited, so we can't always used it as a default. But the macOS (or
really MachO) support was much farther along, so we should be able to
already make the switch. We can likely remove the conditionals once we
update to LLVM 15.